folder페이지에서 현재 유저의 폴더 목록 데이터를 받아올 때 accessToken을 넘기면서 /api/folders 에서 받아옴
